Handle multiple station URLs on Register
Games like Mario Kart 7 or Mario Tennis Open set two station URLs on the
Register method, the local sattion and the public station. In these
cases, use the public station given by the client.

Also add the PID and RVCID on both stations, as they are required on NAT
traversal and not all games will call `ReportNATProperties` to add them.