Redesign find sessions using SearchCriteria
The `MatchmakeSessionSearchCriteria` is intended to be compared with the
original session to find valid sessions. Thus, we can remove the
`SearchCriteria` field that we are storing internally and instead
compare values with the session.

With this, we implement comparing attributes by default, unless
game-specific checks are defined, in which case the attribute search
gets overriden. This allows overriding region-locked matchmaking.

This also allows removing session creation that requires search criteria
`CreateSessionBySearchCriteria` and instead move everything to
`CreateSessionByMatchmakeSession`.

We can also remove the cleanup of search criteria, since it's redundant
considering we allow game-specific checks.

Friends-only matches are also now supported, and requires setting the
`GetUserFriendPIDs` handler on `MatchmakeExtension`.

This improves accuracy finding sessions and is how Rambo's MK8 server is
implemented.

Tested successfully with:
- Mario Kart 7
- Steel Diver: Sub Wars
- IRONFALL Invasion