feat!(ranking): Add new methods for different modes on GetRanking
There's 5 different types of GetRanking call, asking for:
- global leaderboards
- global leaderboards, nearby to your own rank
- only your friends
- only your friends, nearby to your own rank
- your own ranking only

These are pretty fundamentally different requests that would need different
db queries, etc. at the implementation level. Notably all but the first
need knowledge of the PID of the user making the request (to get their rank
and their friends).

Add new methods to the protocol for each of these use cases.