fix(matchmake-extension): Handle switch gathering properly
We were previously removing the participants from their previous
gatherings before creating or joining a new one. However, this is
incorrect, as it was demonstrated by the "AddToGathering" notification
that a user can be in two matchmake sessions at the same time.

Remove this check to make the user persist on both gatherings.